requirements.txt
fastapi
uvicorn
openai
python-multipart
pydantic
gtts
python-dotenv
langdetect